requirements.txt
boto3==1.18.64
botocore==1.21.64
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
jira==3.0.1
jmespath==0.10.0
keyring==23.2.1
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.4
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.25
urllib3==1.26.7
zipp==3.6.0